Move all writer subclasses out of headers
They're now defined in anonymous namespaces in the C++ files, which cuts
down the amount of stuff that has to be sucked into the header files.
